% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_bloodtype.R
\name{em_bloodtype}
\alias{em_bloodtype}
\title{EM algorithm for ABO blood type}
\usage{
em_bloodtype(
  nA,
  nAB,
  nB,
  nO,
  pA0,
  pB0,
  pO0,
  iter = 1000,
  epsilon = 0.01,
  delta = 0.01
)
}
\arguments{
\item{nA}{fraction of A blood type}

\item{nAB}{fraction of AB blood type}

\item{nB}{fraction of B blood type}

\item{nO}{fraction of O blood type}

\item{iter}{max number of iterations}

\item{epsilon}{stopping criteria for allele frequencies}

\item{delta}{stopping criteria for observed likelihood}

\item{PA}{initial guess for A allele frequency}

\item{pB}{initial guess for B allele frequency}

\item{pO}{initial guess for O allele frequency}
}
\value{
population allele frequencies and observed data likelihood
}
\description{
EM algorithm for ABO blood type
}
