% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_casino_baumwelch.R
\name{hmm_casino_baumwelch}
\alias{hmm_casino_baumwelch}
\title{HMM parameter estimation using forward and backward subsitution}
\usage{
hmm_casino_baumwelch(e, x, y, iter = 1000, eps = 0.05)
}
\arguments{
\item{x}{hidden states}

\item{y}{observations}

\item{iter}{max iterations}

\item{E}{emission probabilities for fair die}
}
\value{
all unkown parameters
}
\description{
HMM parameter estimation using forward and backward subsitution
}
