% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_casino_forwardbackward.R
\name{hmm_casino_forwardbackward}
\alias{hmm_casino_forwardbackward}
\title{HMM parameter estimation using forward and backward subsitution}
\usage{
hmm_casino_forwardbackward(P, E, v, y)
}
\arguments{
\item{P}{transition probabilities}

\item{E}{emission probabilities}

\item{v}{initial distribution}

\item{y}{observations}
}
\value{
forward and backward probs, and hidden state marginal probs
}
\description{
HMM parameter estimation using forward and backward subsitution
}
