% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_casino_simulate.R
\name{hmm_casino_simulate}
\alias{hmm_casino_simulate}
\title{HMM simulation occasional dishonest casino HMM}
\usage{
hmm_casino_simulate(P, E, v, n = 100)
}
\arguments{
\item{P}{transition probabilities}

\item{E}{emission probabilities}

\item{v}{initial distribution}

\item{n}{number of simulated points}
}
\value{
simulated values
}
\description{
HMM simulation occasional dishonest casino HMM
}
