% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_fit.R
\name{logreg_fit}
\alias{logreg_fit}
\title{fit logistic regression model to data}
\usage{
logreg_fit(
  X,
  y,
  method = "gradient descent",
  iter = 1000,
  stepsize = 0.001,
  tol = 0.01
)
}
\arguments{
\item{X}{data}

\item{y}{labels}

\item{method}{optimization method, by default gradient descent}

\item{iter}{max number of iterations}

\item{stepsize}{gradient descent stepsize}

\item{tol}{stopping criteria for log likelihoods}
}
\value{
m model coefficients,
}
\description{
fit logistic regression model to data
}
