% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_mult.R
\name{mat_mult}
\alias{mat_mult}
\title{Multiple two matrices and a vector}
\usage{
mat_mult(A, B, x, left = TRUE)
}
\arguments{
\item{A}{a square matrix}

\item{B}{a square matrix with the same dim as A}

\item{x}{a vector with length equal to length of A}

\item{left}{a boolean option that determines which product is computed first}
}
\value{
matrix product \eqn{A \times B \times x}
}
\description{
Multiple two matrices and a vector
}
\examples{
mat_mult(cbind(c(1,2),c(3,4)),rbind(c(1,2),(3,4)),c(1,2),True)
}
