% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings_beta.R
\name{metropolis_hastings_beta}
\alias{metropolis_hastings_beta}
\title{metropolis hastings for beta distribution wiht log normal porpo}
\usage{
metropolis_hastings_beta(a = 4, b = 2, N = 1000, sigma = 1)
}
\arguments{
\item{a}{parameter for beta}

\item{b}{parameter for beta}

\item{N}{length of markov chain}

\item{sigma}{for log normal proposal density}
}
\value{
markov chain
}
\description{
metropolis hastings for beta distribution wiht log normal porpo
}
