% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_chol.R
\name{mvn_chol}
\alias{mvn_chol}
\title{Simulation of MVN random vectors}
\usage{
mvn_chol(mu, sigma, N)
}
\arguments{
\item{mu}{n-dim vector, mean}

\item{sigma}{nxn positive definite matrix, covariance matrix}

\item{N}{number of relizations from MVN(mu,sigma)}
}
\value{
N realizations from MVN distribution
}
\description{
Simulation of MVN random vectors
}
\examples{
mat_mult(cbind(c(1,2),c(3,4)),rbind(c(1,2),(3,4)),c(1,2),True)
}
