% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posdef.R
\name{posdef}
\alias{posdef}
\title{Generate random positive definite matrix}
\usage{
posdef(n, ev = runif(n, 0, 10))
}
\arguments{
\item{n}{dimension}

\item{ev}{eigenvalues (optional)}
}
\value{
positive definite matrix
}
\description{
From R Varadhan (https://stat.ethz.ch/pipermail/r-help/2008-February/153708)
}
