% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_qr.R
\name{reg_qr}
\alias{reg_qr}
\title{Linear regression with QR decomposition}
\usage{
reg_qr(x, y)
}
\arguments{
\item{y}{response}

\item{X}{covariates}
}
\value{
regresion coefficients
}
\description{
Linear regression with QR decomposition
}
