% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_svd.R
\name{reg_svd}
\alias{reg_svd}
\title{Linear regression with SVD decomposition}
\usage{
reg_svd(x, y)
}
\arguments{
\item{y}{response}

\item{X}{covariates}
}
\value{
regresion coefficients
}
\description{
Linear regression with SVD decomposition
}
